requirements.txt
numpy>=1.21.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
setuptools>=60.0.0
